Migrate proxy.js to TS, use strict types
