+++ src/-private/create-store.ts
@@ -1,12 +1,13 @@
 import {
-  createStore as reduxCreateStore,
-  Store,
-  Reducer,
   Action,
-  StoreEnhancer,
+  createStore as reduxCreateStore,
   PreloadedState,
+  Reducer,
+  Store,
+  StoreEnhancer,
 } from 'redux';
 import { createNode, updateNode } from './proxy';
+import { Node } from './tracking';
 
 export default function createStore<S, A extends Action, Ext, StateExt>(
   reducer: Reducer<S, A>,
@@ -30,6 +31,7 @@ export default function createStore<S, A extends Action, Ext, StateExt>(
     enhancer
   );
 
+  let rootNode: Node<{ state: S }>;
   const originalGetState = store.getState.bind(store);
   const ensureRootNode = (): void => {
     if (rootNode === undefined) {
@@ -37,9 +39,7 @@ export default function createStore<S, A extends Action, Ext, StateExt>(
         state: originalGetState(),
       });
     }
-  }
-
-  let rootNode;
+  };
 
   store.getState = (): S => {
     ensureRootNode();
