+++ src/index.ts
@@ -1,2 +1,3 @@
+import './-private/devtools';
 export * from 'redux';
 export { default as createStore } from './-private/create-store';
